# speed up dynamic bucket addition