Remove the need to iterate over every bucket on each new dynamic bucket.

Before (admittedly contrived since each bucket is new):

```
$ go test -bench .
testing: warning: no tests to run
BenchmarkDynamicBucket-4           10000            237119 ns/op
PASS
ok      github.com/maniksurtani/quotaservice/benchmark  2.385s
```

After:

```
$ go test -bench .
testing: warning: no tests to run
BenchmarkDynamicBucket-4          500000              3459 ns/op
BenchmarkFindBucket-4           30000000                58.5 ns/op
PASS
ok      github.com/maniksurtani/quotaservice/benchmark  4.495s
```